+++ test/conftest.py
@@ -54,6 +54,30 @@ def load_simple_plugin(qfield_bot: "QFieldBot", data_path: "Path"):
 
 
 @pytest.fixture
+def load_simple_plugin_without_project_loaded_signal(
+    qfield_bot: "QFieldBot", data_path: "Path"
+):
+    qfield_bot.load_plugin(
+        data_path / "simple_plugin" / "main.qml", emit_load_project_ended=False
+    )
+    qfield_bot.show_window()
+    assert qfield_bot.iface.logged_messages == []
+
+
+@pytest.fixture
+def project_loaded(qfield_bot: "QFieldBot", qfield_project_file: "Path"):
+    qfield_bot.open_project(qfield_project_file)
+
+
+@pytest.fixture
+def qfield_project_file(tmp_path: Path, data_path: Path) -> Path:
+    copied_project = get_copied_path(tmp_path, data_path / "qfield_project")
+    project_file = copied_project / "pytest-qfield-project_qfield.qgs"
+    assert project_file.exists()
+    return project_file
+
+
+@pytest.fixture
 def gpkg(tmp_path: Path, data_path: Path) -> Path:
     return get_copied_gpkg(tmp_path, data_path)
 
@@ -86,10 +110,16 @@ def layer_points(gpkg: Path):
 
 
 def get_copied_gpkg(tmp_path: Path, data_path: Path) -> Path:
-    db = data_path / "db.gpkg"
-    new_db_path = tmp_path / "db.gpkg"
-    shutil.copy(db, new_db_path)
-    return new_db_path
+    return get_copied_path(tmp_path, data_path / "qgis_project" / "db.gpkg")
+
+
+def get_copied_path(tmp_path: Path, path: Path) -> Path:
+    new_path = tmp_path / path.name
+    if path.is_dir():
+        shutil.copytree(path, new_path)
+    else:
+        shutil.copy(path, new_path)
+    return new_path
 
 
 def get_gpkg_layer(name: str, gpkg: Path) -> QgsVectorLayer: