+++ test/test_plugin.py
@@ -15,7 +15,6 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with pytest-qfield.  If not, see <https://www.gnu.org/licenses/>.
-
 from typing import TYPE_CHECKING
 
 import pytest
@@ -26,6 +25,7 @@ if TYPE_CHECKING:
     from pytest_subtests import SubTests
 
     from pytest_qfield.qfieldbot import QFieldBot
+    from pytest_qfield.stub_interface.qgis_stubs import QgsProjectStub
 
 
 @pytest.mark.usefixtures("load_simple_plugin")
@@ -44,3 +44,17 @@ def test_qfield_bot_should_load_plugin(
             "{random-uuid-value}",
         ]
         assert qfield_bot.iface.toast_messages == ["Toast displayed!"]
+
+
+@pytest.mark.usefixtures(
+    "load_simple_plugin_without_project_loaded_signal", "project_loaded"
+)
+def test_qfield_bot_should_open_project_file(
+    qfield_bot: "QFieldBot", qgs_project_stub: "QgsProjectStub"
+):
+    assert qfield_bot.iface.logged_messages == ["Project load ended"]
+    assert len(qgs_project_stub.mapLayersByName("points")) == 1
+    osm_layers = qgs_project_stub.mapLayersByName("OpenStreetMap")
+    assert osm_layers, "Expected 'OpenStreetMap' layer in project"
+    osm_layer = osm_layers[0]
+    assert osm_layer.isValid, "OpenStreetMap layer is invalid"