+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+- feat: Functionality to open projects
+- feat: Stub class for QgsRasterLayer
+
 ## 0.1.0 - 2026-03-24
 
 - Initial release